+++ ui/packages/catalog-ui-search/src/main/webapp/component/component.less
@@ -29,9 +29,9 @@
 @import 'editor/editor';
 @import 'editor/metacard-basic/metacard-basic';
 @import 'filter/filter';
-@import 'filter-search-form/filter';
+@import 'filter/search-form/filter.search-form';
 @import 'filter-builder/filter-builder';
-@import 'filter-builder-search-form/filter-builder';
+@import 'filter-builder/search-form/filter-builder.search-form';
 @import 'filter-comparator/filter-comparator';
 @import 'workspaces/workspaces';
 @import 'workspaces-items/workspaces-items';
@@ -85,7 +85,7 @@
 @import 'multivalue/multivalue';
 @import 'query-annotations/query-annotations';
 @import 'query-advanced/query-advanced';
-@import 'query-custom/query-custom';
+@import 'query-advanced/query-custom/query-custom';
 @import 'query-item/query-item';
 @import 'query-settings/query-settings';
 @import 'query-src/query-src';
@@ -223,7 +223,7 @@
 @import 'search-settings/search-settings';
 @import 'search-form/search-form.less';
 @import 'search-form-selector/search-form-selector.less';
-@import 'dropdown/search-type/dropdown.search-type.less';
+@import 'dropdown/search-form-selector/dropdown.search-form-selector.less';
 @import 'workspace-lists/workspace-lists';
 @import 'dropdown/list-select/dropdown.list-select.less';
 @import 'list-select/list-select';
