+++ ui/packages/catalog-ui-search/src/main/webapp/component/dropdown/search-form-selector/dropdown.search-form-selector.hbs
