+++ ui/packages/catalog-ui-search/src/main/webapp/component/dropdown/search-form-selector/dropdown.search-form-selector.less
@@ -1,4 +1,4 @@
-@{customElementNamespace}dropdown.is-search-type {
+@{customElementNamespace}dropdown.is-search-form-selector {
     cursor: pointer;
     display: block;
