+++ ui/packages/catalog-ui-search/src/main/webapp/component/dropdown/search-form-selector/dropdown.search-form-selector.view.js
@@ -18,14 +18,14 @@ define([
     'underscore',
     'jquery',
     '../dropdown.view',
-    './dropdown.search-type.hbs',
+    './dropdown.search-form-selector.hbs',
     'component/search-form-selector/search-form-selector.view',
     'js/store'
 ], function (Marionette, _, $, DropdownView, template, SearchForms, store) {
 
     return DropdownView.extend({
         template: template,
-        className: 'is-search-type',
+        className: 'is-search-form-selector',
         componentToShow: SearchForms,
         initialize: function(){
             DropdownView.prototype.initialize.call(this);
