+++ ui/packages/catalog-ui-search/src/main/webapp/component/filter-builder/search-form/filter-builder.search-form.hbs
