+++ ui/packages/catalog-ui-search/src/main/webapp/component/filter-builder/search-form/filter-builder.search-form.less
@@ -1,0 +1,3 @@
+@{customElementNamespace}filter-builder.is-search-form {
+
+}
