+++ ui/packages/catalog-ui-search/src/main/webapp/component/filter/search-form/filter.search-form.hbs
