+++ ui/packages/catalog-ui-search/src/main/webapp/component/filter/search-form/filter.search-form.less
@@ -1,0 +1,3 @@
+@{customElementNamespace}filter.is-search-form {
+
+}
