+++ ui/packages/catalog-ui-search/src/main/webapp/component/query-add/query-add.view.js
@@ -17,7 +17,7 @@ var Marionette = require('marionette');
 var template = require('./query-add.hbs');
 var CustomElements = require('js/CustomElements');
 var QueryBasic = require('component/query-basic/query-basic.view');
-var QueryCustom = require('component/query-custom/query-custom.view');
+var QueryCustom = require('component/query-advanced/query-custom/query-custom.view');
 var QueryAdvanced = require('component/query-advanced/query-advanced.view');
 var QueryTitle = require('component/query-title/query-title.view');
 var QueryAdhoc = require('component/query-adhoc/query-adhoc.view');
