+++ ui/packages/catalog-ui-search/src/main/webapp/component/query-advanced/query-custom/query-custom.less
@@ -1,4 +1,4 @@
-@{customElementNamespace}query-custom {
+@{customElementNamespace}query-advanced.is-search {
   .customElement;
   overflow: hidden;
