+++ ui/packages/catalog-ui-search/src/main/webapp/component/search-form-selector/search-form-selector.less
@@ -1,6 +1,6 @@
 @{customElementNamespace}search-form-selector {
 
     .title {
-        padding:10px;
+        padding:@minimumSpacing;
     }
 }
