+++ ui/packages/catalog-ui-search/src/main/webapp/component/search-form/search-form.less
@@ -29,5 +29,5 @@
 }
 
 @{customElementNamespace}search-form-loading {
-    padding: 2*@minimumSpacing;
+    padding: 2 * @minimumSpacing;
 }
