--- ui/packages/catalog-ui-search/src/main/webapp/component/search-type/search-type.less
@@ -1,3 +1,0 @@
-@{customElementNamespace}search-type {
-    .customElement;
-}
