+++ ui/packages/catalog-ui-search/src/main/webapp/component/tabs/search-form/tabs.search-form.js
